+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+## [0.5.0] - 2023-04-17
 ### Changed
 - **Breaking Changed**: `format` and `write` take `Context` trait, this breaks type inference
 - `i` allows access to outer variables
@@ -42,7 +43,8 @@ Fixed URLs in Cargo.toml
 ## [0.1.0] - 2023-02-27
 Initial release
 
-[unreleased]: https://github.com/ModProg/interpolator/compare/v0.4.0...HEAD
+[unreleased]: https://github.com/ModProg/interpolator/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/ModProg/interpolator/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/ModProg/interpolator/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/ModProg/interpolator/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/ModProg/interpolator/compare/v0.1.2...v0.2.0